Remove struct template `TransitionAssertion`
