--- external/concealer/include/concealer/transition_assertion.hpp
@@ -1,68 +1,0 @@
-// Copyright 2015 TIER IV, Inc. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef CONCEALER__TRANSITION_ASSERTION_HPP_
-#define CONCEALER__TRANSITION_ASSERTION_HPP_
-
-#include <chrono>
-#include <concealer/get_parameter.hpp>
-#include <rclcpp/node.hpp>
-#include <rclcpp/rate.hpp>
-#include <scenario_simulator_exception/exception.hpp>
-
-namespace concealer
-{
-template <typename Autoware>
-struct TransitionAssertion
-{
-protected:
-  std::chrono::steady_clock::time_point start;
-
-  std::chrono::seconds initialize_duration;
-
-  bool engaged = false;
-
-  explicit TransitionAssertion()
-  : start(std::chrono::steady_clock::now()),
-    initialize_duration(getParameter<int>("initialize_duration"))
-  {
-  }
-
-  template <typename Thunk = void (*)(), typename Interval = std::chrono::seconds>
-  auto waitForAutowareStateToBe(
-    const std::string & state, Thunk thunk = [] {}, Interval interval = std::chrono::seconds(1))
-  {
-    for (thunk(); not static_cast<const Autoware &>(*this).is_stop_requested.load() and
-                  static_cast<const Autoware &>(*this).autoware_state != state;
-         rclcpp::GenericRate<std::chrono::steady_clock>(interval).sleep()) {
-      if (not engaged and start + initialize_duration <= std::chrono::steady_clock::now()) {
-        throw common::AutowareError(
-          "Simulator waited for the Autoware state to transition to ", state,
-          ", but time is up. The current Autoware state is ",
-          (static_cast<const Autoware &>(*this).autoware_state.empty()
-             ? "not published yet"
-             : static_cast<const Autoware &>(*this).autoware_state));
-      } else {
-        thunk();
-      }
-    }
-
-    if (state == "DRIVING") {
-      engaged = true;
-    }
-  }
-};
-}  // namespace concealer
-
-#endif  // CONCEALER__TRANSITION_ASSERTION_HPP_
